Refactor: read pseudopotentials in esolver_fp
